release 1.3.1: decodeable winget exit codes + timezone env
  internal/winget                runWinget2 sign-extends the Windows
                                 HRESULT-style exit code that Go's
                                 ExitCode() returns as uint32, so the
                                 lookup table keyed on Microsoft's signed
                                 int32 constants (-1978...) actually
                                 matches. Expanded the known-code table:
                                 package_already_installed (0x8A15002C),
                                 installer_not_applicable (0x8A15002E),
                                 update_not_applicable (0x8A150033),
                                 update_all_has_failure, install_failed,
                                 dependency_not_found, download_failed,
                                 operation_canceled, package_agreements_
                                 not_accepted (0x8A150111) and more.
                                 Unknown negatives now render as
                                 exit:0x8A15XXXX (matches Microsoft's
                                 returnCodes doc format).
  internal/winupdate             Same int32 normalisation in
                                 runPSStreaming for consistency.

  .env.example                   New TZ var, default UTC, documented as
                                 an IANA zone name.
  docker-compose.yml             Thread TZ to BOTH services — MariaDB's
                                 time_zone=SYSTEM follows the container
                                 TZ, Go's time.Now() follows the server
                                 container TZ. CURRENT_TIMESTAMP and
                                 formatted timestamps now agree.
  docker-compose.build.yml       Same.
  DOCKERHUB.md                   Compose snippet + env table updated.

Smoke-tested on the docker host with TZ=Europe/Bratislava: both
containers report CEST, MariaDB NOW() returns local time, UTC_TIMESTAMP()
returns UTC with the correct 2h offset.

+++ internal/winupdate/winupdate.go
@@ -129,7 +129,10 @@ func runPSStreaming(ctx context.Context, script string, logf func(string, ...any
 	code := 0
 	if err != nil {
 		if ee, ok := err.(*exec.ExitError); ok {
-			code = ee.ExitCode()
+			// Normalize to signed int32 so HRESULT-style codes
+			// (should a COM call surface one through PS) match the same
+			// format used by winget.ExitCodeResult.
+			code = int(int32(ee.ExitCode()))
 		} else {
 			return allOutput.String(), -1
 		}
